% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{addSeven}
\alias{addSeven}
\title{Add seven}
\usage{
addSeven(x)
}
\arguments{
\item{x}{numeric. A single number or a vector with data.}
}
\value{
A single number or a vector with 7 added to each of its elements.
}
\description{
Adds seven to any number or vector given to the function.
}
\examples{
# add 7 to a vector of numbers
x <- c(3, 4, 5, 6, 7)
addSeven(x)
# add 7 to a single number
addSeven(3)
}
\references{
Sarafoglou, A. (2024). How to add 7 to a number. Science, 10, 206-208.
}
