% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\docType{data}
\name{numbers}
\alias{numbers}
\title{Numbers}
\format{
A data frame with 100 rows and 3 variables:
\describe{
\item{natural_numbers}{numeric. Natural numbers ranging from 1 to 100}
\item{integers}{numeric. Integers ranging from -49 to 50.}
\item{real_numbers}{numeric. Real numbers drawn from a normal distribution with mean 0 and standard deviation of 30.}
}
}
\source{
Sarafoglou, A. (2024). How to add 7 to a number. Science, 10, 206-208.
}
\usage{
numbers
}
\description{
A data frame which features natural numbers, integers, and real numbers
}
\examples{
data(numbers)
}
\keyword{datasets}
